# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Notifications API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "jamshedzodnekruz@gmail.com"
        },
        "version": "1.0"
    },
    "host": "api-notifications.dev.my.cloud",
    "basePath": "/api",
    "paths": {
        "/notifications-external/v1/push": {
            "post": {
                "security": [
                    {
                        "SignatureAuth": []
                    }
                ],
                "description": "All fields except `personExternalRef` (crm_client_id) are required.\n- If you want to send push with `personExternalRef`, do not provide `phone`.\n- If `showInFeed` is true, the push will be shown in the feed; otherwise, it will be hidden.\n- If the users status is inactive or their push setting is disabled, the push will be saved in the feed but not sent to the device.\nIn that case, the payload will be `inactive_user#fake_message_id` or `disabled_push#fake_message_id`.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "External"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Provide user ID created on the server side",
                        "name": "X-UserId",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Provide unique request ID to build hash and track the request",
                        "name": "X-RequestId",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Provide current date in RFC1123 format (e.g., Mon, 02 Jan 2006 15:04:05 MST) to build hash",
                        "name": "X-Date",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Provide user action (push, sms) to send push",
                        "name": "X-UserAction",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Provide hash sum built with HMAC-SHA256 from the `X-Date:X-RequestId` using the secret key created on the server side",
                        "name": "X-RequestDigest",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Request payload",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/push.externalRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/resp.Response"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/resp.Response"
                        }
                    },
                    "401": {
                        "description": "Invalid authorization data",
                        "schema": {
                            "$ref": "#/definitions/resp.Response"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/resp.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Error",
                        "schema": {
                            "$ref": "#/definitions/resp.Response"
                        }
                    }
                }
            }
        },
...
        "/notifications-internal/v1/events/{id}/run": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Events"
                ],
                "summary": "Run event manually",
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/resp.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "payload": {
                                            "$ref": "#/definitions/event.eventModel"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/resp.Response"
                        }
                    },
                    "401": {
                        "description": "Invalid authorization data",
                        "schema": {
                            "$ref": "#/definitions/resp.Response"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/resp.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Error",
                        "schema": {
                            "$ref": "#/definitions/resp.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "event.eventModel": {
            "type": "object",
            "properties": {
                "body": {
                    "$ref": "#/definitions/language.Language"
                },
                "category": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "extraData": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "image": {
                    "$ref": "#/definitions/language.Language"
                },
                "link": {
                    "type": "string"
                },
                "scheduledAt": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "title": {
                    "$ref": "#/definitions/language.Language"
                },
                "topic": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "language.Language": {
            "type": "object",
            "properties": {
                "en": {
                    "type": "string"
                },
                "ru": {
                    "type": "string"
                },
                "tg": {
                    "type": "string"
                },
                "uz": {
                    "type": "string"
                }
            }
        },
        "push.externalRequest": {
            "type": "object",
            "required": [
                "body",
                "phone",
                "showInFeed",
                "title",
                "type"
            ],
            "properties": {
                "body": {
                    "$ref": "#/definitions/language.Language"
                },
                "personExternalRef": {
                    "type": "string",
                    "example": "123456"
                },
                "phone": {
                    "type": "string",
                    "example": "+992111111111"
                },
                "showInFeed": {
                    "type": "boolean"
                },
                "title": {
                    "$ref": "#/definitions/language.Language"
                },
                "type": {
                    "type": "string",
                    "example": "otp, push"
                }
            }
        },
        "resp.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "payload": {}
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}